refactor: add license and rename components
