+++ packages/apps/stock/src/components/templates/customerDelivery/index.js
@@ -1,8 +1,26 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2023 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 export {default as CustomerDeliveryCard} from './CustomerDeliveryCard/CustomerDeliveryCard';
-export {default as CustomerDeliveryDetailHeader} from './CustomerDeliveryDetailHeader/CustomerDeliveryDetailHeader';
-export {default as CustomerDeliveryDetailMovementIndicationCard} from './CustomerDeliveryDetailMovementIndicationCard/CustomerDeliveryDetailMovementIndicationCard';
-export {default as CustomerDeliveryDetailViewAllContainer} from './CustomerDeliveryDetailViewAllContainer/CustomerDeliveryDetailViewAllContainer';
+export {default as CustomerDeliveryHeader} from './CustomerDeliveryHeader/CustomerDeliveryHeader';
+export {default as CustomerDeliveryMovementIndicationCard} from './CustomerDeliveryMovementIndicationCard/CustomerDeliveryMovementIndicationCard';
+export {default as CustomerDeliveryViewAllContainer} from './CustomerDeliveryViewAllContainer/CustomerDeliveryViewAllContainer';
 export {default as CustomerDeliveryLineCard} from './CustomerDeliveryLineCard/CustomerDeliveryLineCard';
-export {default as CustomerDeliveryLineDetailFixedItems} from './CustomerDeliveryLineDetailFixedItems/CustomerDeliveryLineDetailFixedItems';
+export {default as CustomerDeliveryLineButtons} from './CustomerDeliveryLineButtons/CustomerDeliveryLineButtons';
 export {default as CustomerDeliveryLineQuantityCard} from './CustomerDeliveryLineQuantityCard/CustomerDeliveryLineQuantityCard';
 export {default as CustomerDeliveryRealizeButton} from './CustomerDeliveryRealizeButton/CustomerDeliveryRealizeButton';
