+++ packages/apps/stock/src/components/templates/internalMove/index.js
@@ -1,3 +1,21 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2023 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 export {default as InternalMoveCard} from './InternalMoveCard/InternalMoveCard';
 export {default as InternalMoveLineButtons} from './InternalMoveLineButtons/InternalMoveLineButtons';
 export {default as InternalMoveLineCard} from './InternalMoveLineCard/InternalMoveLineCard';
